Refactor core js files
